requirements.txt
datasets==2.19.1
huggingface-hub==0.23.2
pandas==2.2.2
pinecone_client==4.1.0
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2